added an endpoint mapper
